kimvieware-phase0-validator/requirements.txt
-e ../kimvieware-shared
python-magic==0.4.27
